server/requirements.txt
torch
torchvision
Pillow
efficientnet-pytorch
matplotlib
grpcio
grpcio-tools